Resolve "[SCRIPT] Create a set of events from an input image"

Closes #9

See merge request gscarpellini/event-library!11

+++ event_library/representations/constant_count.py
@@ -8,8 +8,9 @@ from matplotlib import pyplot as plt
 
 
 def get_generator(
-    events: np.array, num_events: int, frame_size: Tuple[int, int]) -> np.array:
-    event_count_frame = np.zeros((frame_size[0], frame_size[1], 1))
+    events: np.array, num_events: int, frame_size: Tuple[int, int]
+) -> np.array:
+    event_count_frame = np.zeros((frame_size[0], frame_size[1], 1), dtype="int")
 
     for ind, event in enumerate(events):
         y = int(event[0])
